+++ crates/mongodb-agent-common/src/query/serialization/tests.rs
@@ -1,7 +1,7 @@
 use configuration::MongoScalarType;
 use mongodb::bson::Bson;
 use mongodb_cli_plugin::type_from_bson;
-use mongodb_support::BsonScalarType;
+use mongodb_support::{BsonScalarType, ExtendedJsonMode};
 use ndc_query_plan::{self as plan, inline_object_types};
 use plan::QueryContext;
 use proptest::prelude::*;
@@ -19,7 +19,9 @@ proptest! {
         let inferred_type = inline_object_types(&object_types, &inferred_schema_type.into(), MongoConfiguration::lookup_scalar_type)?;
         let error_context = |msg: &str, source: String| TestCaseError::fail(format!("{msg}: {source}\ninferred type: {inferred_type:?}\nobject types: {object_types:?}"));
 
-        let json = bson_to_json(&inferred_type, bson.clone()).map_err(|e| error_context("error converting bson to json", e.to_string()))?;
+        // Test using Canonical mode because Relaxed mode loses some information, and so does not
+        // round-trip precisely.
+        let json = bson_to_json(ExtendedJsonMode::Canonical, &inferred_type, bson.clone()).map_err(|e| error_context("error converting bson to json", e.to_string()))?;
         let actual = json_to_bson(&inferred_type, json.clone()).map_err(|e| error_context("error converting json to bson", e.to_string()))?;
         prop_assert!(custom_eq(&actual, &bson),
             "`(left == right)`\nleft: `{:?}`\nright: `{:?}`\ninferred type: {:?}\nobject types: {:?}\njson_representation: {}",
@@ -37,7 +39,7 @@ proptest! {
     fn converts_datetime_from_bson_to_json_and_back(d in arb_datetime()) {
         let t = plan::Type::Scalar(MongoScalarType::Bson(BsonScalarType::Date));
         let bson = Bson::DateTime(d);
-        let json = bson_to_json(&t, bson.clone())?;
+        let json = bson_to_json(ExtendedJsonMode::Canonical, &t, bson.clone())?;
         let actual = json_to_bson(&t, json.clone())?;
         prop_assert_eq!(actual, bson, "json representation: {}", json)
     }