+++ crates/mongodb-support/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod align;
 mod bson_type;
 pub mod error;
+mod extended_json_mode;
 
 pub use self::bson_type::{BsonScalarType, BsonType};
+pub use self::extended_json_mode::ExtendedJsonMode;
 
 pub const EXTENDED_JSON_TYPE_NAME: &str = "ExtendedJSON";