+++ crates/ndc-query-plan/src/plan_for_query_request/mod.rs
@@ -17,9 +17,10 @@ use indexmap::IndexMap;
 use itertools::Itertools;
 use ndc::{ExistsInCollection, QueryRequest};
 use ndc_models as ndc;
+use query_plan_state::QueryPlanInfo;
 
 use self::{
-    helpers::{find_object_field, lookup_relationship},
+    helpers::{find_object_field, find_object_field_path, lookup_relationship},
     query_context::QueryContext,
     query_plan_error::QueryPlanError,
     query_plan_state::QueryPlanState,
@@ -42,14 +43,38 @@ pub fn plan_for_query_request<T: QueryContext>(
     )?;
     query.scope = Some(Scope::Root);
 
-    let unrelated_collections = plan_state.into_unrelated_collections();
+    let QueryPlanInfo {
+        unrelated_joins,
+        variable_types,
+    } = plan_state.into_query_plan_info();
+
+    // If there are variables that don't have corresponding entries in the variable_types map that
+    // means that those variables were not observed in the query. Filter them out because we don't
+    // need them, and we don't want users to have to deal with variables with unknown types.
+    let variables = request.variables.map(|variable_sets| {
+        variable_sets
+            .into_iter()
+            .map(|variable_set| {
+                variable_set
+                    .into_iter()
+                    .filter(|(var_name, _)| {
+                        variable_types
+                            .get(var_name)
+                            .map(|types| !types.is_empty())
+                            .unwrap_or(false)
+                    })
+                    .collect()
+            })
+            .collect()
+    });
 
     Ok(QueryPlan {
         collection: request.collection,
         arguments: request.arguments,
         query,
-        variables: request.variables,
-        unrelated_collections,
+        variables,
+        variable_types,
+        unrelated_collections: unrelated_joins,
     })
 }
 
@@ -478,11 +503,11 @@ fn plan_for_binary_comparison<T: QueryContext>(
         plan_for_comparison_target(plan_state, root_collection_object_type, object_type, column)?;
     let (operator, operator_definition) = plan_state
         .context
-        .find_comparison_operator(comparison_target.get_column_type(), &operator)?;
+        .find_comparison_operator(comparison_target.get_field_type(), &operator)?;
     let value_type = match operator_definition {
-        plan::ComparisonOperatorDefinition::Equal => comparison_target.get_column_type().clone(),
+        plan::ComparisonOperatorDefinition::Equal => comparison_target.get_field_type().clone(),
         plan::ComparisonOperatorDefinition::In => {
-            plan::Type::ArrayOf(Box::new(comparison_target.get_column_type().clone()))
+            plan::Type::ArrayOf(Box::new(comparison_target.get_field_type().clone()))
         }
         plan::ComparisonOperatorDefinition::Custom { argument_type } => argument_type.clone(),
     };
@@ -519,20 +544,20 @@ fn plan_for_comparison_target<T: QueryContext>(
                 path,
                 requested_columns,
             )?;
-            let column_type = find_object_field(&target_object_type, &name)?.clone();
+            let field_type = find_object_field_path(&target_object_type, &name, &field_path)?.clone();
             Ok(plan::ComparisonTarget::Column {
                 name,
                 field_path,
                 path,
-                column_type,
+                field_type,
             })
         }
         ndc::ComparisonTarget::RootCollectionColumn { name, field_path } => {
-            let column_type = find_object_field(root_collection_object_type, &name)?.clone();
+            let field_type = find_object_field_path(root_collection_object_type, &name, &field_path)?.clone();
             Ok(plan::ComparisonTarget::ColumnInScope {
                 name,
                 field_path,
-                column_type,
+                field_type,
                 scope: plan_state.scope.clone(),
             })
         }
@@ -559,10 +584,13 @@ fn plan_for_comparison_value<T: QueryContext>(
             value,
             value_type: expected_type,
         }),
-        ndc::ComparisonValue::Variable { name } => Ok(plan::ComparisonValue::Variable {
-            name,
-            variable_type: expected_type,
-        }),
+        ndc::ComparisonValue::Variable { name } => {
+            plan_state.register_variable_use(&name, expected_type.clone());
+            Ok(plan::ComparisonValue::Variable {
+                name,
+                variable_type: expected_type,
+            })
+        }
     }
 }
 
@@ -603,7 +631,7 @@ fn plan_for_exists<T: QueryContext>(
                             comparison_target.column_name().to_owned(),
                             plan::Field::Column {
                                 column: comparison_target.column_name().to_string(),
-                                column_type: comparison_target.get_column_type().clone(),
+                                column_type: comparison_target.get_field_type().clone(),
                                 fields: None,
                             },
                         )